feat: implemented custom exceptions

+++ LeaveManagmentClean.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using LeaveManagmentClean.Application.Contracts.Persistence;
+using LeaveManagmentClean.Application.Exceptions;
 using MediatR;
 
 namespace LeaveManagmentClean.Application.Features.LeaveType.Commands.DeleteLeaveType
@@ -16,7 +17,12 @@ namespace LeaveManagmentClean.Application.Features.LeaveType.Commands.DeleteLeav
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
-
+            
+            if (leaveTypeToDelete is null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
+            
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
 
             return Unit.Value;
